ファイルわけ

+++ src/app/api/blog/[id]/route.ts
@@ -1,55 +1,66 @@
 import { NextResponse } from "next/server";
-import main from "../route";
 import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
+import createClient from "../../libs";
 
 //ブログの詳細記事API
 export const GET = async (req: Request, res: NextResponse) => {
+  let prisma;
   try {
     const id: number = parseInt(req.url.split("/blog/")[1]);
-    await main();
-    const post = await prisma.post.findFirst({ where: { id } });
-    return NextResponse.json({ message: "Success", post }, { status: 200 });
+    prisma = await createClient();
+    if (prisma) {
+      const post = await prisma.post.findFirst({ where: { id } });
+      return NextResponse.json({ message: "Success", post }, { status: 200 });
+    }
   } catch (err) {
     return NextResponse.json({ message: "error", err }, { status: 500 });
   } finally {
-    await prisma.$disconnect();
+    if (prisma) {
+      await prisma.$disconnect();
+    }
   }
 };
 
 //ブログ記事の編集
 export const PUT = async (req: Request, res: NextResponse) => {
-    try {
-      const id: number = parseInt(req.url.split("/blog/")[1]);
-      const {title,description}=await req.json();
-      await main();
-
+  let prisma;
+  try {
+    const id: number = parseInt(req.url.split("/blog/")[1]);
+    const { title, description } = await req.json();
+    prisma = await createClient();
+    if (prisma) {
       const post = await prisma.post.update({
-        data:{title,description},
-        where:{id},
+        data: { title, description },
+        where: { id },
       });
       return NextResponse.json({ message: "Success", post }, { status: 200 });
-    } catch (err) {
-      return NextResponse.json({ message: "error", err }, { status: 500 });
-    } finally {
+    }
+  } catch (err) {
+    return NextResponse.json({ message: "error", err }, { status: 500 });
+  } finally {
+    if (prisma) {
       await prisma.$disconnect();
     }
-  };
+  }
+};
 
-  //削除
-  export const DELETE = async (req: Request, res: NextResponse) => {
-    try {
-      const id: number = parseInt(req.url.split("/blog/")[1]);
-      await main();
-
+//削除
+export const DELETE = async (req: Request, res: NextResponse) => {
+  let prisma;
+  try {
+    const id: number = parseInt(req.url.split("/blog/")[1]);
+    prisma = await createClient();
+    if (prisma) {
       const post = await prisma.post.delete({
-        where:{id},
+        where: { id },
       });
       return NextResponse.json({ message: "Success", post }, { status: 200 });
-    } catch (err) {
-      return NextResponse.json({ message: "error", err }, { status: 500 });
-    } finally {
+    }
+  } catch (err) {
+    return NextResponse.json({ message: "error", err }, { status: 500 });
+  } finally {
+    if (prisma) {
       await prisma.$disconnect();
     }
-  };
+  }
+};
